+++ app/src/main/java/opensource/cached_dupe_scanner/cache/CacheDatabaseFactory.kt
@@ -33,7 +33,8 @@ val CACHE_DATABASE_MIGRATIONS: Array<Migration> = arrayOf(
     CacheMigrations.MIGRATION_21_22,
     CacheMigrations.MIGRATION_22_23,
     CacheMigrations.MIGRATION_23_24,
-    CacheMigrations.MIGRATION_24_25
+    CacheMigrations.MIGRATION_24_25,
+    CacheMigrations.MIGRATION_25_26
 )
 
 fun buildCacheDatabase(